+++ CHANGELOG.rst
@@ -19,6 +19,7 @@ Added
 - Add ``RestClient.disseminate`` method for making new DIPs to v3 rest api client.
 - Add separate classes for transfer ID, AIP ID and DIP ID to v3 rest api client.
 - Add ``list_aip_files`` method for getting file information from AIP to v3 rest api client.
+- Add ``list_aip_divs`` method for getting information about AIP's structmap divs to v3 rest api client.
 
 Changed
 ~~~~~~~